♻️ Refactor: MainArea부분 컴포넌트화
